+++ src/dom7.js
@@ -1,10 +1,10 @@
 import Dom7 from './dom7-class';
 import $ from './$';
 import Utils from './utils';
-import Ajax from './ajax';
+import { ajax, get, post, getJSON } from './ajax';
 import Scroll from './scroll';
 import Methods from './methods';
-import { Animate, Stop } from './animate';
+import { animate, stop } from './animate';
 
 function dom7() {
   // Utils & Helpers
@@ -23,43 +23,14 @@ function dom7() {
   });
 
   // Animate
-  Dom7.prototype.animate = Animate;
-  Dom7.prototype.stop = Stop;
+  Dom7.prototype.animate = animate;
+  Dom7.prototype.stop = stop;
 
   // Ajax
-  $.ajax = Ajax;
-
-  // Ajax Shrotcuts
-  ('get post getJSON').split(' ').forEach((method) => {
-    $[method] = function ajax(...args) {
-      let url;
-      let data;
-      let success;
-      let error;
-      let dataType;
-      if (typeof args[1] === 'function') {
-        [url, success, error, dataType] = args;
-      } else {
-        [url, data, success, error, dataType] = args;
-      }
-      [success, error].forEach((callback) => {
-        if (typeof callback === 'string') {
-          dataType = callback;
-          if (callback === success) success = undefined;
-          else error = undefined;
-        }
-      });
-      dataType = dataType || (method === 'getJSON' ? 'json' : undefined);
-      return $.ajax({
-        url,
-        method: method === 'post' ? 'POST' : 'GET',
-        data,
-        success,
-        error,
-        dataType,
-      });
-    };
-  });
+  $.ajax = ajax;
+  $.get = get;
+  $.post = post;
+  $.getJSON = getJSON;
 
   // Link to prototype
   $.fn = Dom7.prototype;